# Optimized fused MoE Kernel, take 2