--- a/benchmarks/kernels/benchmark_mixtral_moe.py
+++ b/benchmarks/kernels/benchmark_mixtral_moe.py
@@ -12,7 +12,7 @@ import triton
 
 def main():
     method = fused_moe
-    for bs in [1, 2, 4, 8, 16, 32, 64, 128, 1024, 2048, 4096, 8192]:
+    for bs in [1, 2, 4, 8, 16, 24, 32, 48, 64, 96, 128, 256, 512, 1024, 1536, 2048, 3072, 4096]:
         run_grid(bs, method=method)