--- a/vllm/model_executor/layers/fused_moe/__init__.py
+++ b/vllm/model_executor/layers/fused_moe/__init__.py
@@ -1,1 +1,6 @@
 from vllm.model_executor.layers.fused_moe.fused_moe import fused_moe, get_moe_configs
+
+__all__ = [
+    "fused_moe",
+    "get_moe_configs",
+]