--- a/vllm/model_executor/layers/fused_moe/__init__.py
+++ b/vllm/model_executor/layers/fused_moe/__init__.py
@@ -1,6 +1,5 @@
-from vllm.model_executor.layers.fused_moe.fused_moe import fused_moe, get_moe_configs
+from vllm.model_executor.layers.fused_moe.fused_moe import fused_moe
 
 __all__ = [
     "fused_moe",
-    "get_moe_configs",
 ]