--- a/vllm/model_executor/layers/fused_moe/fused_moe.py
+++ b/vllm/model_executor/layers/fused_moe/fused_moe.py
@@ -1,5 +1,7 @@
 """Fused MoE kernel."""
 import os
+from typing import Dict, Any
+
 import torch
 import triton
 import triton.language as tl