This replaces https://github.com/vllm-project/vllm/pull/2913

With some more aggressive parameter search, we were actually able to beat the TensorRT kernels even on small batch sizes, which is very fortunate since it reduces the complexity a lot.

Here are the results on this PR on H100 with TP2:

```
qps = 1 => 11.6 ms ITL
qps = 2 => 12.6 ms ITL
qps = 4 => 23.6 ms ITL
qps = 6 => 32.9 ms ITL
```

current main branch (untuned fused MoE kernel):

```
qps = 1 => 23.3 ms ITL
qps = 2 => 25.4ms ITL
qps = 4 => 43.0ms ITL
qps = 6 => 60.8ms ITL
```

only using the TensorRT Moe kernels:

```
qps = 1 => 18.1 ms ITL
qps = 2 => 23.8 ms ITL
qps = 4 => 48.1 ms ITL
qps = 6 => 90.8 ms ITL
```

The code is structured in such a way that it is very easy to just drop a new .json file into the `fused_moe/configs` directory to support a different configuration.

Co-authored-by: Cade Daniel <edacih@gmail.com>