@pcmoritz 2024-02-24
@WoosukKwon The PR only includes optimized settings for TP2 on H100 (and I also added the optimizations for TP4 on A100 80 GB) so you can't measure a difference for TP4 on H100. I added a README to explain this. For Mixtral it only really makes sense to run TP2 on H100. Instead of TP4 on H100, it is better to run two replicas of TP2 on H100 and split the traffic between them. Similarly, on A100, TP4 is the optimal setting in my experience.