@WoosukKwon 2024-02-24
@pcmoritz Got it. Thanks for the explanation! BTW, how did you tune the knobs? I believe the performance ranking between different configurations is also dependent on the number of input tokens. What batch size did you tune for?