@njhill 2024-02-26
This is awesome thanks @pcmoritz! TP2 A100 would also be very useful, we can have a go at adding a config for that.