@paolovic 2024-04-07
Do I understand it correctly, that I can use this PR to improve inference time on my cluster? If so, how?
I have Mixtral-8x7B-Instruct-v0.1-GPTQ (8bit-128g-actorder_True) deployed on  2x Nvidia L40S